pyproject.toml
[tool.poetry]
name = "media-platform"
version = "0.1.0"
description = ""
authors = ["Damian Ontivero <damian.ontivero@011h.com>"]
readme = "README.md"
packages = [{include = "media_platform", from = "src"}]

[[tool.poetry.source]]
name = "test"
url = "https://test.pypi.org/simple/"
priority = "supplemental"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.13.1"
cryptography = "^42.0.5"
fastapi = "^0.110.0"
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.29"
uvicorn = "^0.29.0"
werkzeug = "^3.0.1"
ditainer = "^0.2.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
factory-boy = "^3.3.0"
flake8 = "^7.0.0"
ipdb = "^0.13.13"
isort = "^5.13.2"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 79
